Fixed a minor bug in removeRedundantPhases.
Since the minimization function used for removeRedundantPhases could be
slightly different than the function used in traceMinimum, it can happen that
the minimization in the former can drop into a different phase entirely. This
should only happen at either the beginning or the end of a phase (where the
phase disappears and the minimization can be unstable), so the fix does a
simple check to see if the minimization is being done at the end points, and
if it is, it just uses the end point instead of redoing the minimization.