switch from `optimize.fmin` to `optimize.fminbound` in `findRScale`
There were rare circumstances where `optimize.fmin` would yield
a value outside of the expected range even though the input to
the minimized function was clamped and there was in fact a minimum
within the expected range. `fminbound` is a much more natural function
to use here, and it doesn't seem to suffer from the same problem.